All commands should now use new argument parsing
