+++ src/main/kotlin/me/bristermitten/devdenbot/commands/arguments/Arguments.kt
@@ -1,0 +1,37 @@
+package me.bristermitten.devdenbot.commands.arguments
+
+data class Arguments(val command: String, val args: List<Argument>, val flags: List<Argument>) {
+
+    inline fun validateArgLength(length: Int, orElse: () -> Unit) {
+        if (args.size < length)
+            orElse()
+    }
+
+}
+
+
+class Argument(rawContent: String) {
+    val isFlag = rawContent.startsWith('-')
+    val content = rawContent.removePrefix("-")
+
+    inline fun validate(predicate: (String) -> Boolean, orElse: () -> Unit) {
+        if (predicate(content)) return
+        orElse()
+    }
+
+    override fun equals(other: Any?) =
+        when (other) {
+            is String -> content.equals(other, true)
+            is Argument -> content == other.content && isFlag == other.isFlag
+            else -> false
+        } // Doesn't this break the symmetrical contract?
+
+    override fun hashCode(): Int = content.lowercase().hashCode()
+
+    override fun toString(): String {
+        return "Argument(isFlag=$isFlag, content='$content')"
+    }
+
+
+}
+
